Moved Reputation to a module folder, moved ChatReputationData and SingleReputationData classes to the globals
