+++ lib/src/globals/chat_reputation_data.dart
@@ -1,0 +1,6 @@
+class ChatReputationData {
+  final String name;
+  final int reputation;
+
+  ChatReputationData({required this.name, required this.reputation});
+}
