+++ lib/src/globals/single_reputation_data.dart
@@ -1,0 +1,8 @@
+class SingleReputationData {
+  final String id;
+  final int reputation;
+  final int increaseOptionsLeft;
+  final int decreaseOptionsLeft;
+
+  SingleReputationData({required this.id, required this.reputation, required this.increaseOptionsLeft, required this.decreaseOptionsLeft});
+}
