+++ lib/src/modules/reputation/reputation_manager.dart
@@ -1,8 +1,9 @@
-import 'package:weather/src/modules/chat/chat.dart';
-import 'package:weather/src/core/reputation.dart';
+import 'package:weather/src/globals/chat_reputation_data.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/platform/platform.dart';
-import 'utils.dart';
+import 'package:weather/src/modules/chat/chat.dart';
+import 'reputation.dart';
+import '../utils.dart';
 
 class ReputationManager {
   final Platform platform;
