+++ test/reputation_test.dart
@@ -2,7 +2,7 @@ import 'package:postgres/postgres.dart';
 import 'package:test/test.dart';
 import 'package:weather/src/core/event_bus.dart';
 import 'package:weather/src/events/accordion_poll_events.dart';
-import 'package:weather/src/core/reputation.dart';
+import 'package:weather/src/modules/reputation/reputation.dart';
 import 'package:weather/src/modules/user/user.dart';
 import 'package:weather/src/modules/chat/chat.dart';
 import 'utils/setup.dart';
